requirements.txt
Flask==2.0.2
Pillow==8.4.0
rembg==1.0.29
Flask-MySQLdb==0.2.0
